refactor(seed): extract method bulkUpsert from GPC seeder
