+++ app/seeders/20230927132757-gpc.cjs
@@ -2,6 +2,7 @@
 
 const fs = require("node:fs");
 const { parse } = require("csv-parse");
+const { bulkUpsert } = require("./util/util.cjs");
 
 async function parseFile(filename) {
   const records = [];
@@ -20,34 +21,6 @@ async function parseFile(filename) {
   }
 
   return records;
-}
-
-async function bulkUpsert(
-  queryInterface,
-  tableName,
-  entries,
-  idColumnName,
-  transaction,
-) {
-  for (const entry of entries) {
-    const id = entry[idColumnName];
-    const item = await queryInterface.sequelize.query(
-      `SELECT COUNT(*) FROM "${tableName}" WHERE "${idColumnName}" = '${id}';`,
-      { transaction },
-    );
-    if (item[0][0].count === "0") {
-      await queryInterface.bulkInsert(tableName, [entry], { transaction });
-    } else {
-      await queryInterface.bulkUpdate(
-        tableName,
-        entry,
-        { [idColumnName]: entry[idColumnName] },
-        {
-          transaction,
-        },
-      );
-    }
-  }
 }
 
 /** @type {import('sequelize-cli').Migration} */
